Refactor CloseButton, BotonSolicitar
