+++ DREAM Lab Frontend/src/pages/HomePage/components/Detalles/components/CloseButton/CloseButton.jsx
